requirements.txt
pyTelegramBotAPI==4.6.0
numpy==1.23.0
opencv-python==3.4.8.29
torch==1.11.0
torchvision==0.12.0
pillow==9.1.1
tensorflow==2.9.1
ffmpeg-python==0.2.0